+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.4] - 2025-12-27
+
+### Added
+- Amazon metadata provider settings (enable/disable, domain selection)
+- Comic Vine metadata provider settings (enable/disable, API key)
+
 ## [0.4.3] - 2025-12-27
 
 ### Added